chore: migrate the signup page to use the app router
