+++ apps/web/app/future/signup/page.tsx
@@ -1,0 +1,10 @@
+import Signup from "@pages/signup";
+import { _generateMetadata } from "app/_utils";
+
+export const generateMetadata = async () =>
+  await _generateMetadata(
+    (t) => t("sign_up"),
+    (t) => t("sign_up")
+  );
+
+export default Signup;
