+++ apps/web/lib/buildLegacyCtx.tsx
@@ -51,3 +51,5 @@ export const buildLegacyCtx = (
     }
   ) as unknown as GetServerSidePropsContext;
 };
+
+export type LegacyCtx = ReturnType<typeof buildLegacyCtx>;
